Add a multiple ESM interface mode to iOS ESM plugin.
For using the interface, a user has to add 'interface:1' to the ESM configuration file.
The default value of the 'interface' element is zero (0). By using the zero interface mode, a user can use the normal interface (single ESM). In the future development, if we need to make a new interface mode, we should use interface 2 as a new interface.

This version is a prototype. For public release, we need tests carefully and widely.

[interface]
0 = single ESM interface
1 = multiple ESM interface