Fix errors on iOSESMViewController.m
[error1: unsorted ESMs on multiple-ESM interface]
Unsorted objects appear when for-in is used for NSSet. I used sort method for the NSSet object before using the for-in.

[error2: crash when a user push button in the multiple-ESM interface]
The reason of the error is that the same tag (0) is tagged to all ESM object. Each ESM object should use a different tag. I update the ESM generation method for setting a different tag to each ESM.